requirements/testing.txt
-r base.txt

testfixtures
mock==1.0.1
model-mommy==1.2.2

coverage
coveralls
nose>=1.3.0
django-nose>=1.2
flake8>=2.1.0
tox>=1.7.0

# Additional test requirements go here

beautifulsoup4==4.3.2
requests>=2.3.0
dataset==0.5.5
SQLAlchemy==0.9.8
django-debug-toolbar==1.5
